added gainDao into handler, added access modifier, added gainDao usage into handle method

+++ src/main/java/core/model/Gain.java
@@ -1,13 +1,13 @@
 package core.model;
 
 public class Gain {
-    private int gain;
+    private double gain;
 
-    public Gain(int gain) {
+    public Gain(double gain) {
         this.gain = gain;
     }
 
-    public int getGain() {
+    public double getGain() {
         return gain;
     }
